# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: Fraud Rule Engine API
  description: "Fraud rule engine: deterministic evaluation with\
    \ optional AI advisory. Submit transaction events, retrieve flagged decisions\
    \ with full rule trace."
  version: v1
servers:
- url: http://localhost:8090
  description: Generated server url
tags:
- name: decisions
  description: Retrieve evaluated decisions.
- name: auth
  description: Demo-only token issuance. Production deployment uses an external IdP.
- name: transactions
  description: Submit transaction events for fraud evaluation.
- name: advisory
  description: Optional LLM-backed commentary on a decision. Never authoritative.
- name: audit
  description: Read the audit log. Service API key required.
- name: admin
  description: Operator endpoints. Day-4 work will require service API key.
paths:
  /auth/token:
    post:
      tags:
      - auth
      summary: Issue a short-lived demo JWT. Always with role USER — this endpoint
        cannot mint elevated roles regardless of request body.
      operationId: token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /api/v1/transactions:
    post:
      tags:
      - transactions
      summary: "Submit a transaction event. Persists the raw event, the verdict, the\
        \ matched-rule trace, and an outbox row in a single transaction. Supports\
        \ the standard Idempotency-Key header (24h cache)."
      operationId: submit
      parameters:
      - name: Idempotency-Key
        in: header
        description: Optional. Cached for 24h scoped to the JWT subject. Replay with
          same key + same body returns the cached response; same key + different body
          returns 409 Conflict.
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/rules/reload:
    post:
      tags:
      - admin
      summary: Hot-reload the rule set from the configured YAML resource. Atomic swap
        via AtomicReference; old set serves in-flight evaluations until completion.
        On validation failure the old set stays active and a 422 is returned.
      operationId: reload
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/v1/decisions:
    get:
      tags:
      - decisions
      summary: "Page through recent decisions, newest first."
      operationId: list
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DecisionPage"
  /api/v1/decisions/{id}:
    get:
      tags:
      - decisions
      summary: "Fetch a single decision by id, including the full matched-rule trace."
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DecisionDetail"
  /api/v1/decisions/{id}/advisory:
    get:
      tags:
      - advisory
      summary: Generate or return advisory commentary for the given decision id.
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdvisoryResponse"
  /admin/audit:
    get:
      tags:
      - audit
      summary: "Page through audit log entries, newest first."
      operationId: list_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuditPage"
components:
  schemas:
    TokenRequest:
      required:
      - subject
      type: object
      properties:
        subject:
          maxLength: 128
          minLength: 0
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
    TransactionRequest:
      required:
      - accountAgeDays
      - accountId
      - amount
      - channel
      - country
      - currency
      - ipCountry
      - mcc
      - timestamp
      - txId
      type: object
      properties:
        txId:
          type: string
          format: uuid
        accountId:
          type: string
        amount:
          minimum: 0.0
          exclusiveMinimum: false
          type: number
        currency:
          pattern: "[A-Z]{3}"
          type: string
        mcc:
          pattern: "[0-9]{4}"
          type: string
        channel:
          pattern: WEB|MOBILE|POS|ATM|API
          type: string
        country:
          pattern: "[A-Z]{2}"
          type: string
        ipCountry:
          pattern: "[A-Z]{2}"
          type: string
        deviceId:
          type: string
        merchantId:
          type: string
        accountAgeDays:
          minimum: 0
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
    DecisionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DecisionSummary"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    DecisionSummary:
      type: object
      properties:
        decisionId:
          type: string
          format: uuid
        txId:
          type: string
          format: uuid
        accountId:
          type: string
        status:
          type: string
        score:
          type: number
          format: double
        ruleSetVersion:
          type: integer
          format: int32
        evaluatedAt:
          type: string
          format: date-time
    DecisionDetail:
      type: object
      properties:
        decisionId:
          type: string
          format: uuid
        txId:
          type: string
          format: uuid
        accountId:
          type: string
        status:
          type: string
        score:
          type: number
          format: double
        ruleSetVersion:
          type: integer
          format: int32
        evaluatedAt:
          type: string
          format: date-time
        matchedRules:
          type: array
          items:
            $ref: "#/components/schemas/MatchedRule"
    MatchedRule:
      type: object
      properties:
        ruleId:
          type: string
        priority:
          type: integer
          format: int32
        reason:
          type: string
    AdvisoryResponse:
      type: object
      properties:
        advisoryStatus:
          type: string
          enum:
          - OK
          - TIMED_OUT
          - UNAVAILABLE
          - MALFORMED
        summary:
          type: string
        concerns:
          type: array
          items:
            type: string
        confidence:
          type: number
          format: double
        humanReviewRequired:
          type: boolean
    AuditDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actor:
          type: string
        action:
          type: string
        resourceId:
          type: string
        payloadHash:
          type: string
        occurredAt:
          type: string
          format: date-time
    AuditPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64